+++ fltk/src/macros/browser.rs
@@ -18,10 +18,20 @@ macro_rules! impl_browser_ext {
                     unsafe { [<$flname _add>](self.inner, item.as_ptr()) }
                 }
 
+                fn add_with_data<T: Clone + 'static>(&mut self, item: &str, data: T) {
+                    self.add(item);
+                    self.set_data(self.size(), data);
+                }
+
                 fn insert(&mut self, line: i32, item: &str) {
                     assert!(!self.was_deleted());
                     let item = CString::safe_new(item);
                     unsafe { [<$flname _insert>](self.inner, line as i32, item.as_ptr()) }
+                }
+
+                fn insert_with_data<T: Clone + 'static>(&mut self, line: i32, item: &str, data: T) {
+                    self.insert(line, item);
+                    self.set_data(line, data);
                 }
 
                 fn move_item(&mut self, to: i32, from: i32) {
@@ -313,6 +323,24 @@ macro_rules! impl_browser_ext {
                 fn value(&self) -> i32 {
                     assert!(!self.was_deleted());
                     unsafe { [<$flname _value>](self.inner) as i32 }
+                }
+
+                fn set_data<T: Clone + 'static>(&mut self, line: i32, data: T) {
+                    assert!(!self.was_deleted());
+                    unsafe {
+                        [<$flname _set_data>](self.inner, line, Box::into_raw(Box::from(data)) as _);
+                    }
+                }
+            
+                unsafe fn data<T: Clone + 'static>(&self, line: i32) -> Option<T> {
+                    assert!(!self.was_deleted());
+                    let ptr = [<$flname _data>](self.inner, line);
+                    if ptr.is_null() {
+                        None
+                    } else {
+                        let data = ptr as *const _ as *mut T;
+                        Some((*data).clone())
+                    }
                 }
             }
         }