+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 
-## [1.3.3] - 2022-04-10
+## [1.3.5] - 2022-04-17
+- Add BrowserExt::data(), set_data(), add_with_data() and insert_with_data() methods.
+- Fix SharedImage::from_image() refcount.
+
+## [1.3.4] - 2022-04-10
 - Pull FLTK fixes.
 - Some refactoring.